get rid of cart component
